# src/main/resources/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: MoneyGuard API
  description: |
    # MoneyGuard API
    
    API для приложения управления личными финансами MoneyGuard, которое позволяет пользователям отслеживать доходы и расходы, 
    устанавливать бюджеты, создавать финансовые цели и получать аналитику по своим финансам.
    
    ## Аутентификация
    
    Все защищенные эндпоинты требуют JWT токен, который можно получить через `/auth/login` или `/auth/register`.
    Токен необходимо передавать в заголовке `Authorization` в формате `Bearer {token}`.
    
    ## Доступ к ИИ
    
    Приложение предоставляет базовый функционал для всех пользователей. Доступ к AI-рекомендациям можно получить,
    активировав соответствующую опцию через API подписки.
  version: 1.0.0
  contact:
    email: 123
servers:
  - url: http://localhost:8080
    description: Локальный сервер разработки
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT токен авторизации, полученный при регистрации или входе в систему
  parameters:
...
    goalId:
      name: id
      in: path
      required: true
      schema:
        type: integer
      description: Уникальный идентификатор цели
      example: 1
  schemas:
...
    TransactionResponse:
      type: object
      properties:
        id:
          type: integer
          description: Уникальный идентификатор транзакции
          example: 1
        amount:
          type: number
          description: Сумма транзакции (отрицательная для расходов, положительная для доходов)
          example: -1000
        category:
          type: string
          description: Название категории
          example: Продукты
        date:
          type: string
          format: date
          description: Дата транзакции
          example: 2023-04-01
        description:
          type: string
          description: Описание транзакции
          example: Покупка продуктов в магазине
        goal_id:
          type: integer
          nullable: true
          description: ID цели, к которой относится транзакция
          example: null
        created_at:
          type: string
          format: date-time
          description: Дата создания записи о транзакции
          example: 2023-04-01T12:00:00Z
...
    GoalRequest:
      type: object
      required: [ name, target_amount, target_date, priority ]
      properties:
        name:
          type: string
          description: Название цели
          example: Новый ноутбук
        target_amount:
          type: number
          format: double
          minimum: 0
          description: Целевая сумма
          example: 150000
        target_date:
          type: string
          format: date
          description: Дата достижения цели
          example: 2023-09-01
        priority:
          type: string
          enum: [ LOW, MEDIUM, HIGH ]
          description: Приоритет цели
          example: HIGH
    GoalResponse:
      type: object
      properties:
        id:
          type: integer
          description: Уникальный идентификатор цели
          example: 1
        name:
          type: string
          description: Название финансовой цели
          example: Новый автомобиль
        target_amount:
          type: number
          description: Целевая сумма
          example: 1000000
        current_amount:
          type: number
          description: Текущая накопленная сумма
          example: 150000
        target_date:
          type: string
          format: date
          description: Целевая дата достижения цели
          example: 2024-12-31
        priority:
          type: string
          description: Приоритет цели
          example: HIGH
    GoalPlanResponse:
      type: object
      properties:
        id:
          type: integer
          description: Уникальный идентификатор цели
          example: 1
        name:
          type: string
          description: Название финансовой цели
          example: Новый автомобиль
        description:
          type: string
          description: Описание финансовой цели
          example: Накопить на новый автомобиль
        target_amount:
          type: number
          description: Целевая сумма
          example: 1000000
        current_amount:
          type: number
          description: Текущая накопленная сумма
          example: 150000
        target_date:
          type: string
          format: date
          description: Целевая дата достижения цели
          example: 2024-12-31
        priority:
          type: string
          description: Приоритет цели
          example: HIGH
        days_left:
          type: integer
          description: Количество дней до целевой даты
          example: 365
        daily_contribution:
          type: number
          description: Рекомендуемый ежедневный взнос для достижения цели
          example: 2328.77
        monthly_contribution:
          type: number
          description: Рекомендуемый ежемесячный взнос для достижения цели
          example: 70000
        created_at:
          type: string
          format: date-time
          description: Дата создания цели
          example: 2023-04-01T12:00:00Z
    GoalDetailResponse:
      type: object
      properties:
        goal:
          $ref: '#/components/schemas/GoalResponse'
        transactions:
          type: array
          items:
            $ref: '#/components/schemas/TransactionResponse'
          description: Список транзакций, связанных с целью
        progress:
          type: number
          minimum: 0
          maximum: 100
          description: Процент выполнения цели (0-100)
          example: 15
...
    ErrorResponse:
      type: object
      properties:
        status:
          type: integer
          description: HTTP статус код ошибки
          example: 404
        message:
          type: string
          description: Сообщение об ошибке
          example: Категория с ID 999 не найдена
        timestamp:
          type: string
          format: date-time
          description: Время возникновения ошибки
          example: 2023-04-01T12:00:00Z
        path:
          type: string
          description: Путь запроса, вызвавшего ошибку
          example: /api/categories/999
    ValidationErrorResponse:
      type: object
      properties:
        status:
          type: integer
          description: HTTP статус код ошибки
          example: 400
        message:
          type: string
          description: Общее сообщение об ошибке
          example: Ошибка валидации данных
        timestamp:
          type: string
          format: date-time
          description: Время возникновения ошибки
          example: 2023-04-01T12:00:00Z
        path:
          type: string
          description: Путь запроса, вызвавшего ошибку
          example: /api/categories
        errors:
          type: object
          additionalProperties:
            type: string
          description: Детальные ошибки валидации по полям
          example:
            name: Название категории не может быть пустым
            icon: Иконка должна быть указана
paths:
...
  /goals:
    get:
      tags: [ Goals ]
      summary: Получить список целей
      description: |
        Возвращает список финансовых целей текущего пользователя.
        
        Финансовые цели позволяют планировать накопления на конкретные нужды.
      security:
        - bearerAuth: [ ]
      responses:
        '200':
          description: Список целей
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GoalResponse'
              examples:
                example1:
                  summary: Пример ответа со списком целей
                  value:
                    - id: 1
                      name: Новый автомобиль
                      target_amount: 1000000
                      current_amount: 150000
                      target_date: 2024-12-31
                      priority: HIGH
                    - id: 2
                      name: Отпуск
                      target_amount: 200000
                      current_amount: 50000
                      target_date: 2023-07-01
                      priority: MEDIUM
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags: [ Goals ]
      summary: Создать новую цель
      description: |
        Создает новую финансовую цель для текущего пользователя.
        
        При создании цели автоматически рассчитывается план накопления.
      security:
        - bearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GoalRequest'
            examples:
              example1:
                summary: Пример запроса на создание цели
                value:
                  name: Новый ноутбук
                  description: Накопить на новый ноутбук
                  target_amount: 150000
                  target_date: 2023-09-01
                  priority: MEDIUM
      responses:
        '201':
          description: Цель успешно создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GoalPlanResponse'
              examples:
                example1:
                  summary: Пример ответа с созданной целью
                  value:
                    id: 3
                    name: Новый ноутбук
                    description: Накопить на новый ноутбук
                    target_amount: 150000
                    current_amount: 0
                    target_date: 2023-09-01
                    priority: MEDIUM
                    days_left: 153
                    daily_contribution: 980.39
                    monthly_contribution: 29411.76
                    created_at: 2023-04-01T12:00:00Z
        '400':
          description: Некорректные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /goals/{id}:
    get:
      tags: [ Goals ]
      summary: Получить детальную информацию о цели
      description: |
        Возвращает детальную информацию о цели, включая связанные транзакции и прогресс.
        
        Цель должна принадлежать текущему пользователю.
      security:
        - bearerAuth: [ ]
      parameters:
        - $ref: '#/components/parameters/goalId'
      responses:
        '200':
          description: Детальная информация о цели
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GoalDetailResponse'
              examples:
                example1:
                  summary: Пример ответа с детальной информацией о цели
                  value:
                    goal:
                      id: 1
                      name: Новый автомобиль
                      target_amount: 1000000
                      current_amount: 150000
                      target_date: 2024-12-31
                      priority: HIGH
                    transactions:
                      - id: 5
                        amount: 50000
                        category: Накопления
                        date: 2023-03-01
                        description: Перевод на цель
                        created_at: 2023-03-01T10:00:00Z
                      - id: 8
                        amount: 100000
                        category: Накопления
                        date: 2023-04-01
                        description: Перевод на цель
                        created_at: 2023-04-01T10:00:00Z
                    progress: 15
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Цель не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags: [ Goals ]
      summary: Обновить цель
      description: |
        Обновляет информацию о цели.
        
        Цель должна принадлежать текущему пользователю.
        При обновлении цели автоматически пересчитывается план накопления.
      security:
        - bearerAuth: [ ]
      parameters:
        - $ref: '#/components/parameters/goalId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GoalRequest'
            examples:
              example1:
                summary: Пример запроса на обновление цели
                value:
                  name: Новый ноутбук
                  description: Накопить на новый ноутбук MacBook Pro
                  target_amount: 180000
                  target_date: 2023-10-01
                  priority: HIGH
      responses:
        '200':
          description: Цель успешно обновлена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GoalPlanResponse'
              examples:
                example1:
                  summary: Пример ответа с обновленной целью
                  value:
                    id: 3
                    name: Новый ноутбук
                    description: Накопить на новый ноутбук MacBook Pro
                    target_amount: 180000
                    current_amount: 0
                    target_date: 2023-10-01
                    priority: HIGH
                    days_left: 183
                    daily_contribution: 983.61
                    monthly_contribution: 29508.20
                    created_at: 2023-04-01T12:00:00Z
        '400':
          description: Некорректные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Цель не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags: [ Goals ]
      summary: Удалить цель
      description: |
        Удаляет цель по ее ID.
        
        Цель должна принадлежать текущему пользователю.
        Нельзя удалить цель, которая используется в транзакциях.
      security:
        - bearerAuth: [ ]
      parameters:
        - $ref: '#/components/parameters/goalId'
      responses:
        '204':
          description: Цель успешно удалена
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Цель не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Цель используется в транзакциях
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                example1:
                  summary: Пример ошибки при попытке удаления используемой цели
                  value:
                    status: 409
                    message: Невозможно удалить цель, так как она используется в транзакциях
                    timestamp: 2023-04-01T12:00:00Z
                    path: /goals/1
...